--- a/.github/workflows/test_js.yml
+++ b/.github/workflows/test_js.yml
@@ -1,24 +1,25 @@
 on: [push, pull_request]
-name: Test
+name: Test JS
+permissions: {}
 jobs:
   test:
     strategy:
       matrix:
-        go-version: [1.14.x, 1.15.x, 1.16.x]
+        go-version: [1.18.x,1.19.x]
     runs-on: ubuntu-latest
     steps:
     - name: Install Go
-      uses: actions/setup-go@v1
+      uses: actions/setup-go@v3
       with:
         go-version: ${{ matrix.go-version }}
 
     - name: Install wasmbrowsertest
       run: |
-        go get github.com/agnivade/wasmbrowsertest
+        go install github.com/agnivade/wasmbrowsertest@latest
         mv $HOME/go/bin/wasmbrowsertest $HOME/go/bin/go_js_wasm_exec
 
     - name: Checkout code
-      uses: actions/checkout@v2
+      uses: actions/checkout@v3
 
     - name: Test
       run: go test -exec="$HOME/go/bin/go_js_wasm_exec" ./...